improve performance for Zig Fibonacci legacy benchmark
uses the generic variant for Zig Fibonacci
- allows the use of different integer sizes for new and old benchmark
- new benchmark runs faster with `usize`, old benchmark is faster with `u32`

+++ fibonacci/zig/code.zig
@@ -18,7 +18,7 @@ pub fn main() !void {
     const args = args_cli[1..];
 
     // try different integer sizes (u32/u64/u128/usize) to see their impact on performance
-    const runs = try std.fmt.parseInt(usize, args[0], 0);
+    const runs = try std.fmt.parseInt(u32, args[0], 0);
 
     const fn_fib = fibonacci(@TypeOf(runs));
